requirements.txt
asgiref==3.2.10
astroid==2.4.1
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
dodgy==0.2.1
et-xmlfile==1.0.1
flake8==3.8.4
flake8-polyfill==1.0.2
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
openpyxl==3.0.5
openpyxl-image-loader==1.0.5
parso==0.7.1
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
